% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decode_fit_dfs.R
\name{decode_fit_dfs}
\alias{decode_fit_dfs}
\title{read gps fit data files}
\usage{
decode_fit_dfs(
  fitfilename,
  appendSessionUnits = TRUE,
  frames = c("records", "session", "events"),
  checkconda = TRUE,
  requiredVars
)
}
\arguments{
\item{fitfilename}{name (including file path) of .fit file}

\item{appendSessionUnits}{if FALSE, do not append units to variable names
in session dataframe}

\item{frames}{a vector of frames to return from 
c("records","session","events","hrv")}

\item{checkconda}{logical, check conda environment for packages}

\item{requiredVars}{vector of variable names which will be generated as NA
if they are not already in the fit data records}
}
\value{
a list of three dataframes:  \eqn{session},
    \eqn{records}and \eqn{events}
}
\description{
\code{decode_fit_dfs} reads a gps .fit files using call to python fitdecode
 and put into session, data record, and event dataframes
}
