% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitdecodeR.R
\name{fitdecodeR}
\alias{fitdecodeR}
\title{fitdecodeR: A package for reading garmin .fit files}
\description{
The fitdecodeR package uses reticulate to wrap
  the python fitdecode package, which seems to be kept more
  current than the native R solutions for processing .fit
  files
}
\section{read .fit file}{

  \link{decode_fit_dfs}
}

